Sidebar icons updated
