+++ client-admin/src/ui/Cards/PublishedTransactionCard.jsx
@@ -4,8 +4,8 @@ import Divider from "@material-ui/core/Divider";
 import Button from "@material-ui/core/Button";
 import Grid from "@material-ui/core/Grid";
 import ChevronRightIcon from "@material-ui/icons/ChevronRight";
-import TxArrowIcon from "../icons/TxArrowIcon";
-import EditIcon from "../icons/EditIcon";
+import TxArrowIcon from "../Icons/TxArrowIcon";
+import EditIcon from "../Icons/EditIcon";
 import TxStepper from "../TxStepper";
 
 import { usdFormatter, cryptoFormatter } from "../../util";
