+++ client-admin/src/ui/Cards/TransactionDetailsCard.jsx
@@ -5,7 +5,7 @@ import Button from "@material-ui/core/Button";
 import Grid from "@material-ui/core/Grid";
 import ChevronRightIcon from "@material-ui/icons/ChevronRight";
 import FormatListBulletedIcon from "@material-ui/icons/FormatListBulleted";
-import TxArrowIcon from "../icons/TxArrowIcon";
+import TxArrowIcon from "../Icons/TxArrowIcon";
 import { usdFormatter, cryptoFormatter } from "../../util";
 
 const useStyles = makeStyles((theme) => ({
