+++ client-admin/src/ui/Cards/WalletDetailsCard.jsx
@@ -4,7 +4,7 @@ import Chip from "@material-ui/core/Chip";
 import Button from "@material-ui/core/Button";
 import Grid from "@material-ui/core/Grid";
 import EditIcon from "@material-ui/icons/Edit";
-import CopyIcon from "../icons/CopyIcon";
+import CopyIcon from "../Icons/CopyIcon";
 import AddWallet from "../Dialog/AddWallet";
 import { usdFormatter, cryptoFormatter } from "../../util";
