+++ client-admin/src/ui/Icons/AccountsIcon.jsx
@@ -1,0 +1,13 @@
+import React from "react";
+import SvgIcon from "@material-ui/core/SvgIcon";
+
+export default function PlusIcon(props) {
+  return (
+    <SvgIcon {...props}>
+      <path
+        d="M26.5932386,3 L3.49151367,3 C1.47484182,3.10096906 -0.0825194056,4.81026796 0.00413986958,6.82760538 L0.00413986958,22.784467 C-0.0420421822,23.7547661 0.299664317,24.70358 0.953890835,25.4216335 C1.60811735,26.139687 2.52111691,26.5679895 3.49151367,26.6120724 L26.5932386,26.6120724 C28.5803853,26.4754866 30.0928866,24.7739227 29.9955545,22.784467 L29.9955545,6.82760538 C30.0837125,4.841395 28.5760868,3.14531607 26.5932386,3 Z M10.7384465,6.89565171 C12.6174937,6.89565171 14.1407624,8.41892044 14.1407624,10.2979676 C14.1407624,12.1770148 12.6174937,13.7002835 10.7384465,13.7002835 C8.85939933,13.7002835 7.33613063,12.1770148 7.33613063,10.2979676 C7.33613063,9.39561732 7.69458788,8.53022496 8.33264589,7.89216695 C8.97070389,7.25410895 9.83609626,6.89565171 10.7384465,6.89565171 L10.7384465,6.89565171 Z M16.828592,21.3895221 L4.64830107,21.3895221 C4.19247669,21.3734226 3.77331116,21.1355016 3.52581088,20.7523847 C3.27831061,20.3692678 3.23374046,19.8893515 3.40645577,19.467209 C4.92733818,16.701399 7.85474015,15.0052086 11.0106318,15.0612099 C14.003804,15.0209552 16.7576818,16.6917731 18.1044605,19.3651395 C18.3153576,19.8032529 18.2859949,20.3190555 18.026741,20.730405 C17.7674871,21.1417546 17.3148218,21.3907698 16.828592,21.3895221 L16.828592,21.3895221 Z M26.3040418,19.2970932 L23.0888533,19.2970932 C22.5552799,19.3469311 22.0794784,18.9617585 22.0171238,18.4295026 C22.0793197,17.8992989 22.5583773,17.5190945 23.0888533,17.5789236 L26.3040418,17.5789236 C26.8345178,17.5190945 27.3135753,17.8992989 27.3757713,18.4295026 C27.3134167,18.9617585 26.8376152,19.3469311 26.3040418,19.2970932 L26.3040418,19.2970932 Z M26.8484123,14.9931636 L20.4180353,14.9931636 C19.8875593,15.0529927 19.4085017,14.6727883 19.3463058,14.1425846 C19.3591648,13.8744439 19.4807255,13.6231185 19.6829526,13.4465711 C19.8851797,13.2700237 20.1506126,13.1834974 20.4180353,13.2069477 L26.8654239,13.2069477 C27.3962225,13.1661505 27.866534,13.5468788 27.9371534,14.0745383 C27.9201076,14.342905 27.7942657,14.5926881 27.588742,14.7660987 C27.3832182,14.9395094 27.1158203,15.0215224 26.8484123,14.9931636 Z M26.8484123,10.7062455 L20.4180353,10.7062455 C20.1533722,10.7346455 19.888462,10.6549058 19.6834508,10.4851309 C19.4784396,10.3153561 19.3507216,10.0699504 19.3292942,9.80463181 C19.4000127,9.27922441 19.8732834,8.90361273 20.4010237,8.95405282 L26.8484123,8.95405282 C27.3735043,8.91242879 27.841439,9.28380554 27.9201419,9.80463181 C27.8662612,10.3459434 27.3909589,10.7458009 26.8484123,10.7062455 L26.8484123,10.7062455 Z"
+        id="Shape"
+      ></path>
+    </SvgIcon>
+  );
+}
