+++ client-admin/src/ui/Icons/ConversionIcon.jsx
@@ -1,0 +1,22 @@
+import React from "react";
+import SvgIcon from "@material-ui/core/SvgIcon";
+
+export default function PlusIcon(props) {
+  return (
+    <SvgIcon {...props}>
+      <polyline
+        id="Path-2"
+        stroke="#000000"
+        stroke-width="2"
+        points="2 7.425215 17.5329391 7.425215 12.7273543 1"
+      ></polyline>
+      <polyline
+        id="Path-2"
+        stroke="#000000"
+        stroke-width="2"
+        transform="translate(9.766470, 15.212607) scale(-1, -1) translate(-9.766470, -15.212607) "
+        points="2 18.425215 17.5329391 18.425215 12.7273543 12"
+      ></polyline>
+    </SvgIcon>
+  );
+}
