+++ client-admin/src/ui/Icons/PriceTrackerIcon.jsx
@@ -1,0 +1,21 @@
+import React from "react";
+import SvgIcon from "@material-ui/core/SvgIcon";
+
+export default function PlusIcon(props) {
+  return (
+    <SvgIcon {...props}>
+      <path
+        d="M5.93954082,19.4602041 L0.98494898,19.4602041 C0.596938776,19.4602041 0.268622449,19.7885204 0.268622449,20.1765306 L0.268622449,26.0732143 C0.268622449,26.4612245 0.596938776,26.7895408 0.98494898,26.7895408 L5.93954082,26.7895408 C6.32755102,26.7895408 6.65586735,26.4612245 6.65586735,26.0732143 L6.65586735,20.1765306 C6.65586735,19.7885204 6.32755102,19.4602041 5.93954082,19.4602041 Z"
+        id="Path"
+      ></path>
+      <path
+        d="M15.1622449,13.669898 L10.2076531,13.669898 C9.81964286,13.669898 9.49132653,13.9982143 9.49132653,14.3862245 L9.49132653,26.0732143 C9.49132653,26.4612245 9.81964286,26.7895408 10.2076531,26.7895408 L15.1622449,26.7895408 C15.5502551,26.7895408 15.8785714,26.4612245 15.8785714,26.0732143 L15.8785714,14.3862245 C15.8785714,13.9982143 15.5502551,13.669898 15.1622449,13.669898 Z"
+        id="Path"
+      ></path>
+      <path
+        d="M28.2352041,8.20790816 L22.5045918,0.447704082 C22.2061224,0.0596938776 21.6390306,0.0596938776 21.3405612,0.447704082 L15.609949,8.20790816 C15.2517857,8.68545918 15.609949,9.37193878 16.1770408,9.37193878 L18.7140306,9.37193878 L18.7140306,26.0732143 C18.7140306,26.4612245 19.0423469,26.7895408 19.4303571,26.7895408 L24.384949,26.7895408 C24.7729592,26.7895408 25.1012755,26.4612245 25.1012755,26.0732143 L25.1012755,9.34209184 L27.6382653,9.34209184 C28.265051,9.34209184 28.5933673,8.68545918 28.2352041,8.20790816 Z"
+        id="Path"
+      ></path>
+    </SvgIcon>
+  );
+}
