+++ client-admin/src/ui/Icons/TransactionsIcon.jsx
@@ -1,0 +1,19 @@
+import React from "react";
+import SvgIcon from "@material-ui/core/SvgIcon";
+
+export default function PlusIcon(props) {
+  return (
+    <SvgIcon {...props}>
+      <polygon
+        id="Path"
+        transform="translate(9.385077, 11.856130) rotate(-330.000000) translate(-9.385077, -11.856130) "
+        points="16.6272964 10.6999638 15.1099668 8.09910589 10.9485712 10.5268155 8.48576704 6.3052639 5.31431403 11.8561186 2.14285714 17.4069961 8.53577943 17.3781225 14.9287023 17.3492489 12.4658978 13.1276973"
+      ></polygon>
+      <polygon
+        id="Path"
+        transform="translate(20.429380, 9.139865) rotate(-330.000000) translate(-20.429380, -9.139865) "
+        points="14.6181458 12.8179324 13.1389777 10.1951832 17.3353731 7.82848611 14.9344635 3.57142857 21.3271116 3.63588221 27.7197815 3.70034529 24.467628 9.20432429 21.2154725 14.7083021 18.814563 10.4512448"
+      ></polygon>
+    </SvgIcon>
+  );
+}
