--- client-admin/src/ui/icons/CancelIcons.jsx
@@ -1,20 +1,0 @@
-import React from "react";
-import SvgIcon from "@material-ui/core/SvgIcon";
-
-export default function PlusIcon(props) {
-  return (
-    <SvgIcon {...props}>
-      <g
-        id="noun_cancel_2419034"
-        transform="translate(2.000000, 2.000000)"
-        fill="#000000"
-        fill-rule="nonzero"
-      >
-        <path
-          d="M0.914191255,0 C0.680651293,0 0.443937515,0.0906957528 0.266203818,0.268385391 C-0.0887346062,0.623235829 -0.0887346062,1.20918856 0.266203818,1.56377458 L6.70402513,8 L0.266203818,14.4362254 C-0.0887346062,14.7910759 -0.0887346062,15.3770286 0.266203818,15.7316146 C0.438118853,15.903487 0.678270931,16 0.914191255,16 C1.15011158,16 1.38999917,15.9037514 1.56217869,15.7316146 L8,9.29565361 L14.4378213,15.7316146 C14.6097363,15.903487 14.8498884,16 15.0858087,16 C15.3177618,16 15.5505083,15.914857 15.7337962,15.7316146 C16.0887346,15.3767642 16.0887346,14.7908114 15.7337962,14.4362254 L9.29623936,8 L15.7337962,1.56377458 C16.0887346,1.20892414 16.0887346,0.62297141 15.7337962,0.268385391 C15.3788578,-0.086200628 14.7927597,-0.0864650471 14.4380858,0.268385391 L8.00026448,6.70461081 L1.56244318,0.268385391 C1.38497396,0.0909601719 1.14826019,0 0.91445574,0 L0.914191255,0 Z"
-          id="Path"
-        ></path>
-      </g>
-    </SvgIcon>
-  );
-}
