--- client-admin/src/ui/icons/ChevronRightIcon.jsx
@@ -1,24 +1,0 @@
-import React from "react";
-import SvgIcon from "@material-ui/core/SvgIcon";
-
-export default function PlusIcon(props) {
-  return (
-    <SvgIcon {...props}>
-      <g
-        id="Icon/chevron/black"
-        stroke="none"
-        stroke-width="1"
-        fill="none"
-        fill-rule="evenodd"
-      >
-        <path
-          d="M8.60450258,10.0228839 L13.2148012,5.41258532 C13.4986483,5.12873819 13.4990571,4.66951418 13.2160686,4.38652556 C12.931107,4.101564 12.4736973,4.10410434 12.1900088,4.38779287 L7.06858124,9.50922043 C7.06836997,9.5094317 7.06815885,9.50964308 7.06794787,9.50985456 C7.06773656,9.51006536 7.06752519,9.51027648 7.06731392,9.51048775 C6.92504515,9.65275652 6.85443812,9.83800909 6.85470141,10.0231672 C6.85538511,10.2087849 6.92603019,10.3939964 7.06731392,10.5352802 C7.06752519,10.5354914 7.06773656,10.5357026 7.06794805,10.5359135 C7.06815885,10.5361249 7.06836997,10.5363362 7.06858124,10.5365475 L12.1900088,15.6579751 C12.4738559,15.9418222 12.9330799,15.942231 13.2160686,15.6592424 C13.5010301,15.3742808 13.4984898,14.9168712 13.2148012,14.6331826 L8.60450258,10.0228839 Z"
-          id="Path"
-          fill="#000000"
-          fill-rule="nonzero"
-          transform="translate(10.141691, 10.022543) rotate(-180.000000) translate(-10.141691, -10.022543) "
-        ></path>
-      </g>
-    </SvgIcon>
-  );
-}
