--- client-admin/src/ui/icons/PlusIcon.jsx
@@ -1,39 +1,0 @@
-import React from "react";
-import SvgIcon from "@material-ui/core/SvgIcon";
-
-export default function PlusIcon(props) {
-  return (
-    <SvgIcon {...props}>
-      <g
-        id="Icon/add/black"
-        stroke="none"
-        stroke-width="1"
-        fill="none"
-        fill-rule="evenodd"
-        stroke-linecap="round"
-      >
-        <g
-          id="Group"
-          transform="translate(3.000000, 3.000000)"
-          stroke="#000000"
-          stroke-width="1.5"
-        >
-          <line
-            x1="7.08333333"
-            y1="0.416666667"
-            x2="7.08333333"
-            y2="14.5833333"
-            id="Line-2"
-          ></line>
-          <line
-            x1="-3.64153152e-14"
-            y1="7.5"
-            x2="14.1666667"
-            y2="7.5"
-            id="Line-2"
-          ></line>
-        </g>
-      </g>
-    </SvgIcon>
-  );
-}
