--- client-admin/src/ui/icons/ReceiveIcon.jsx
@@ -1,35 +1,0 @@
-import React from "react";
-import SvgIcon from "@material-ui/core/SvgIcon";
-
-export default function PlusIcon(props) {
-  return (
-    <SvgIcon {...props}>
-      <g
-        id="Icon/receive/black"
-        stroke="none"
-        stroke-width="1"
-        fill="none"
-        fill-rule="evenodd"
-        stroke-linecap="round"
-        stroke-linejoin="round"
-      >
-        <g
-          id="Group-4"
-          transform="translate(1.000000, 1.000000)"
-          stroke="#000000"
-          stroke-width="2"
-        >
-          <g id="Group-2" transform="translate(0.000000, 3.000000)">
-            <polyline
-              id="Path-31"
-              transform="translate(14.089844, 7.000000) rotate(-90.000000) translate(-14.089844, -7.000000) "
-              points="21.0898438 3 14.0898438 11 7.08984375 3"
-            ></polyline>
-            <line x1="18.0898437" y1="7" x2="0.5" y2="7" id="Path-32"></line>
-          </g>
-          <line x1="24" y1="0.5" x2="24" y2="18.5" id="Line-8"></line>
-        </g>
-      </g>
-    </SvgIcon>
-  );
-}
