--- client-admin/src/ui/icons/SettingsIcon.jsx
@@ -1,23 +1,0 @@
-import React from "react";
-import SvgIcon from "@material-ui/core/SvgIcon";
-
-export default function PlusIcon(props) {
-  return (
-    <SvgIcon {...props}>
-      <g
-        id="Icon/setting/black"
-        stroke="none"
-        stroke-width="1"
-        fill="none"
-        fill-rule="evenodd"
-      >
-        <path
-          d="M29.906127,13.3347536 C29.8586866,12.9130362 29.3669806,12.5961342 28.9418029,12.5961342 C27.567147,12.5961342 26.3473144,11.7889774 25.8357393,10.5406805 C25.3131134,9.26213337 25.6501078,7.77105318 26.6745975,6.83128642 C26.9970807,6.53648608 27.0362609,6.04299417 26.7657947,5.70019529 C26.0622255,4.80675271 25.2625475,3.9997075 24.3894205,3.30026833 C24.0475146,3.02589529 23.5458741,3.06407087 23.2493994,3.39235855 C22.3552871,4.38257957 20.7492336,4.75060558 19.5081924,4.23277948 C18.2166968,3.68950299 17.4022844,2.38081725 17.4819843,0.976022801 C17.508216,0.534771114 17.1857328,0.151117675 16.7458206,0.0998820261 C15.6253338,-0.0297140294 14.495024,-0.0337325118 13.3711885,0.0909520651 C12.9362994,0.138950604 12.6138162,0.513562457 12.628439,0.949232918 C12.6773305,2.34018593 11.8530952,3.62598865 10.5744364,4.14973085 C9.34824118,4.65059002 7.75346182,4.28568949 6.86113549,3.30428681 C6.56622349,2.98102223 6.07273158,2.94128391 5.72915133,3.20907111 C4.83012752,3.91442638 4.01247799,4.72214133 3.30243449,5.60821666 C3.02549408,5.95268993 3.06623703,6.45198634 3.39195735,6.74834942 C4.43598138,7.69380902 4.7730874,9.19783763 4.23070391,10.4911192 C3.71287781,11.7241235 2.43254469,12.5187784 0.966803261,12.5187784 C0.491171225,12.5034859 0.152390839,12.822732 0.100373818,13.2551653 C-0.0315663518,14.3821264 -0.033129095,15.5304077 0.0938995964,16.6665219 C0.141005139,17.0900253 0.64766879,17.4041367 1.07753477,17.4041367 C2.38376478,17.370761 3.63786614,18.1794805 4.16395245,19.4597021 C4.6883644,20.7382492 4.35125838,22.2284364 3.32520589,23.1689845 C3.00428543,23.4637848 2.96354248,23.9564954 3.23400867,24.2992942 C3.9309921,25.187044 4.73078171,25.9948706 5.60703411,26.7001142 C5.95072596,26.9770546 6.45080379,26.937986 6.74872961,26.6096984 C7.64619067,25.6170216 9.25213245,25.2496653 10.4882623,25.7684961 C11.782995,26.3100982 12.5974074,27.6186723 12.5177075,29.0242481 C12.491699,29.465723 12.8157449,29.8500462 13.2538711,29.9005005 C13.827063,29.9673636 14.4037152,30.0007393 14.9819301,30.0007393 C15.5307879,30.0007393 16.0797572,29.9706007 16.6286149,29.9095421 C17.0636156,29.8615435 17.3858756,29.4869317 17.3712527,29.0504799 C17.3209101,27.6603082 18.1465966,26.3745055 19.423581,25.8516563 C20.6580364,25.3474484 22.2461183,25.716479 23.1385563,26.6968771 C23.4351426,27.0194719 23.9251742,27.0584289 24.2706521,26.7915347 C25.1680015,26.0878538 25.9839767,25.2808086 26.6973689,24.3922774 C26.9741977,24.0485856 26.9351291,23.5485078 26.6077344,23.2522563 C25.5637104,22.3067967 25.22493,20.8025449 25.7673134,19.5101563 C26.2772142,18.2935608 27.5099953,17.4766926 28.8359828,17.4766926 L29.0215028,17.4814925 C29.451592,17.516431 29.8473009,17.1851294 29.8994295,16.7461102 C30.031593,15.6181445 30.0331557,14.4708679 29.906127,13.3347536 Z M15.0237893,20.0355729 C12.2646547,20.0355729 10.0203323,17.7912505 10.0203323,15.0321159 C10.0203323,12.2730929 12.2646547,10.0286589 15.0237893,10.0286589 C17.7828123,10.0286589 20.0271347,12.2730929 20.0271347,15.0321159 C20.0271347,17.7912505 17.7828123,20.0355729 15.0237893,20.0355729 Z"
-          id="Shape"
-          fill="#4D4D4D"
-          fill-rule="nonzero"
-        ></path>
-      </g>
-    </SvgIcon>
-  );
-}
