--- client-admin/src/ui/icons/WalletsIcon.jsx
@@ -1,32 +1,0 @@
-import React from "react";
-import SvgIcon from "@material-ui/core/SvgIcon";
-
-export default function PlusIcon(props) {
-  return (
-    <SvgIcon {...props}>
-      <g
-        id="Icon/my-wallets/black"
-        stroke="none"
-        stroke-width="1"
-        fill="none"
-        fill-rule="evenodd"
-      >
-        <g
-          id="Group"
-          transform="translate(0.000000, 4.000000)"
-          fill="#4D4D4D"
-          fill-rule="nonzero"
-        >
-          <path
-            d="M14.9147727,12.7840909 L14.9147727,8.52272727 C14.9147727,7.05255682 16.1079545,5.859375 17.578125,5.859375 L27.6988636,5.859375 L27.6988636,2.13068182 C27.6988636,0.953480114 26.7453835,8.17124146e-14 25.5681818,8.17124146e-14 L2.13068182,8.17124146e-14 C0.953480114,8.17124146e-14 5.32907052e-14,0.953480114 5.32907052e-14,2.13068182 L5.32907052e-14,19.1761364 C5.32907052e-14,20.3533381 0.953480114,21.3068182 2.13068182,21.3068182 L25.5681818,21.3068182 C26.7453835,21.3068182 27.6988636,20.3533381 27.6988636,19.1761364 L27.6988636,15.4474432 L17.578125,15.4474432 C16.1079545,15.4474432 14.9147727,14.2542614 14.9147727,12.7840909 Z"
-            id="Path"
-          ></path>
-          <path
-            d="M28.2315341,6.92471591 L27.6988636,6.92471591 L17.578125,6.92471591 C16.6992188,6.92471591 15.9801136,7.64382102 15.9801136,8.52272727 L15.9801136,12.7840909 C15.9801136,13.6629972 16.6992188,14.3821023 17.578125,14.3821023 L27.6988636,14.3821023 L28.2315341,14.3821023 C29.1104403,14.3821023 29.8295455,13.6629972 29.8295455,12.7840909 L29.8295455,8.52272727 C29.8295455,7.64382102 29.1104403,6.92471591 28.2315341,6.92471591 Z M19.7088068,12.2514205 C18.8299006,12.2514205 18.1107955,11.5323153 18.1107955,10.6534091 C18.1107955,9.77450284 18.8299006,9.05539773 19.7088068,9.05539773 C20.5877131,9.05539773 21.3068182,9.77450284 21.3068182,10.6534091 C21.3068182,11.5323153 20.5877131,12.2514205 19.7088068,12.2514205 Z"
-            id="Shape"
-          ></path>
-        </g>
-      </g>
-    </SvgIcon>
-  );
-}
